nutri_scan_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
bcrypt==4.1.1
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
ultralytics==8.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
geopy==2.4.1
pandas==2.1.4
flask-jwt-extended==4.6.0
werkzeug==3.0.1
gunicorn==21.2.0
google-generativeai==0.3.2
groq==0.3.0
